Refactor library package name
