+++ ohs-player-library/src/commonMain/kotlin/dev/ohs/player/library/OhsPlayerLibrary.kt
@@ -1,0 +1,3 @@
+package dev.ohs.player.library
+
+object OhsPlayerLibrary
