--- ohs-player-library/src/commonMain/kotlin/dev/ohs/player/ohsplayerlibrary/OhsPlayerLibrary.kt
@@ -1,3 +1,0 @@
-package dev.ohs.player.ohsplayerlibrary
-
-object OhsPlayerLibrary
