+++ ohs-player-library/src/commonTest/kotlin/dev/ohs/player/library/OhsPlayerLibraryTest.kt
@@ -1,4 +1,4 @@
-package dev.ohs.player.ohsplayerlibrary
+package dev.ohs.player.library
 
 import kotlin.test.Test
 import kotlin.test.assertSame
