+++ ohs-player-reference-app/src/androidMain/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Ohsplayerreferenceclientapp</string>
+    <string name="app_name">OhsPlayerReferenceClientApp</string>
 </resources>
