+++ ohs-player-reference-app/src/webMain/resources/index.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Ohsplayerreferenceclientapp</title>
+    <title>OhsPlayerReferenceClientApp</title>
     <link type="text/css" rel="stylesheet" href="styles.css">
 </head>
 <body style="text-align: center; align-content: center">
